requirements.txt
streamlit==1.31.1 
psycopg2-binary==2.9.9 
python-dotenv==1.0.1
plotly